// src/app/utils/enums.ts
export enum LinkRoutes {
  HOME = '/',
  LOGIN = '/login',
  DASHBOARD = '/dashboard',
}

export enum HttpVerbs {
  GET = 'get',
  POST = 'post',
  PUT = 'put',
  DELETE = 'delete',
  PATCH = 'patch',
}

export enum Role {
  TEACHER = 'Teacher',
  STUDENT = 'Student',
  PRINCIPAL = 'Principal',
}

export enum SideBarCurrent {
  Dashboard = 'Dashboard',
  Teachers = 'Teachers',
  Students = 'Students',
  Exams = 'Exams',
}

export enum Option {
  A = 'A',
  B = 'B',
  C = 'C',
  D = 'D',
}

export enum Action {
  VIEW = 'VIEW',
  EDIT = 'EDIT',
  NEW = 'NEW',
  LIST = 'LIST',
}